pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "drpshadow"
version = "0.0.1"
description = "DRP network simulator"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PyYAML",
  "networkx",
  "numpy",
  "matplotlib",
  "bonsai-network-generator @ git+https://github.com/magnified103/bonsai.git@5be64806",
  "bonsai-network @ git+https://github.com/magnified103/bonsainet.git@93d46941"
]

[project.urls]
Homepage = "https://github.com/magnified103/drpshadow"
Issues = "https://github.com/magnified103/drpshadow/issues"
